requirements.txt
tensorflow>=2.13.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.1.0
seaborn>=0.12.0